tests: New air conditioner example using asyncio
